+++ token-metadata/js/test/update.test.ts
@@ -10,8 +10,8 @@ import {
 import test from 'tape';
 import { amman, InitTransactions, killStuckProcess } from './setup';
 import { Keypair, PublicKey } from '@solana/web3.js';
-import { createAndMintDefaultAsset, createDefaultAsset } from './utils/DigitalAssetManager';
-import { UpdateTestData } from './utils/UpdateTestData';
+import { createAndMintDefaultAsset, createDefaultAsset } from './utils/digital-asset-manager';
+import { UpdateTestData } from './utils/update-test-data';
 import { encode } from '@msgpack/msgpack';
 import { PROGRAM_ID as TOKEN_AUTH_RULES_ID } from '@metaplex-foundation/mpl-token-auth-rules';
 import { spokSamePubkey } from './utils';
