+++ token-metadata/js/test/utils/update-test-data.ts
